js: changes in generation process handling, baking chars into array for better performance
perf(js): in previous version characters were picked randomly from whole
range, and then after character pick event, it was validated to check
whether it fits in filters or not. It had performance variable on
selected parameters, so for only 0-9 filter applied it would be around O(5n), but for all selected chars it would be O(n). Now it's thing of the past, as validation isn't required anymore, prebaked array needs it only once, so it's a lot quicker and cheaper. Characters pickup works similar to previous version, but now it places characters instead of ASCII code.

deletes:
- validateCharacter

replaces:
- pickRandomASCIICharacter with pickRandomCharFromArray
- validRanges with charactersForGenerator

adds:
- pushCharactersFromASCIIRange